fastapi-backend/requirements.txt
 fastapi
uvicorn
SQLAlchemy
pydantic
mysql-connector-python
python-multipart
python-dotenv



PyJWT
alembic
pymysql  # (For MySQL, remove if using SQLite or PostgreSQL)
psycopg2  # (For PostgreSQL, remove if using MySQL or SQLite)
python-multipart